+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.3.0] 2025-11-04 More XWN support
-- Added stress button to character sheet TODO 
+- Added readied and stowed modifiers to character sheet (under features/tweaks)
+- Fixed capt support department bug
 
 ## [2.2.0] 2025-11-03 More AWN support
 
@@ -21,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed issues for ship / mech attack
 - Removed roll formula from item description by default
 - Updated WWN / OSE compendium content
-- Updated NPC import from CSV
+- Updated NPC import from CSV for compendium import
 - Added condition to item description and chat messages. Not yet automated. Behind AWN setting.
 - Shock damage now uses dice string
 - Changed several fields to validate that they are valid dice strings (such as damage, trauma, etc)